+++ back/node_watcher/src/tasks/createReferendumStatus.ts
@@ -5,12 +5,12 @@
 import { ApiPromise } from '@polkadot/api';
 import { BlockNumber, Hash } from '@polkadot/types/interfaces';
 import { logger } from '@polkadot/util';
+import { referendumStatus } from 'src/util/statuses';
 
 import { prisma } from '../generated/prisma-client';
 import {
   NomidotReferendumRawEvent,
   NomidotReferendumStatusUpdate,
-  referendumStatus,
   Task,
 } from './types';
