+++ back/node_watcher/src/util/statuses.ts
@@ -1,0 +1,18 @@
+export const preimageStatus = {
+  NOTED: 'Noted',
+  REAPED: 'Reaped',
+  USED: 'Used',
+};
+
+export const proposalStatus = {
+  PROPOSED: 'Proposed',
+  TABLED: 'Tabled',
+};
+
+export const referendumStatus = {
+  CANCELLED: 'Cancelled',
+  EXECUTED: 'Executed',
+  NOTPASSED: 'NotPassed',
+  PASSED: 'Passed',
+  STARTED: 'Started',
+};
